+++ src/component/icon/icon.tsx
@@ -1,5 +1,6 @@
 import styles from './icon.module.scss'
 import * as react from 'react'
+import { CSSProperties } from 'react'
 import { LucideProps } from 'lucide-react'
 
 export interface Props {
@@ -8,11 +9,17 @@ export interface Props {
     onClick?: () => void
     disabled?: boolean
     ariaLabel?: string
+    color?: string
 }
 
-export function Icon({ icon, size, onClick, disabled, ariaLabel }: Props) {
+export function Icon({ icon, size, onClick, disabled, ariaLabel, color }: Props) {
     const className = onClick ? styles.clickable : styles.icon
     const Icon = icon
+    const inlineStyles: CSSProperties = {}
+
+    if (color !== undefined) {
+        inlineStyles.color = color
+    }
 
     return (
         <span
@@ -20,6 +27,7 @@ export function Icon({ icon, size, onClick, disabled, ariaLabel }: Props) {
             onClick={() => onClick && !disabled && onClick()}
             aria-disabled={disabled || false}
             aria-label={ariaLabel}
+            style={inlineStyles}
         >
             <Icon style={{ width: `var(--icon-${size || 'm'})`, height: `var(--icon-${size || 'm'})` }} />
         </span>